requirements.txt
numpy
pandas
torch
torchvision
matplotlib
seaborn
tqdm
statsmodels
panel
holoviews
altair
lpips
pillow
scikit-learn
yellowbrick
einops
plotly
tensorboard
ipykernel
jupyter_bokeh
nbconvert
tabulate
selenium
kaleido